% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bp_cod}
\alias{bp_cod}
\title{Blood pressure (BP) recording codes}
\format{
\subsection{\code{bp_cod}}{

A data frame with 97 rows and 2 columns:
\describe{
\item{snomed_concept_id}{SNOMED Concept ID}
\item{description}{Description of SNOMED Concept ID}
}
}
}
\source{
<https://digital.nhs.uk/data-and-information/data-collections-and-data-sets/data-collections/quality-and-outcomes-framework-qof >
}
\usage{
bp_cod
}
\description{
SNOMED CT clinical code cluster: 999012731000230108.
}
\keyword{datasets}
